spliting up model protocols
